v1.1

+++ src/calorie/FruitCalorie.java
@@ -1,10 +1,8 @@
 package calorie;
 
-import calorie.observer.Observer;
-
 /**
  * FruitCalorie
  */
-public abstract class FruitCalorie {
-    public abstract float calculerCalories(String name);       
+public abstract class FruitCalorie{
+    public abstract double calculerCalories(String name);       
 }
